+++ lib/ansible/ruby/modules/custom/core/utilities/helper/meta.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+# VALIDATED_CHECKSUM: Isys/3jd2iKpxQYjj30pLNba1cQOHiPhSXIiL4W9Ays=
+# See LICENSE.txt for license
+
+require 'ansible/ruby/modules/generated/core/utilities/helper/meta'
+require 'ansible/ruby/modules/free_form'
+
+module Ansible
+  module Ruby
+    module Modules
+      class Meta
+        include FreeForm
+      end
+    end
+  end
+end