+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `gem` module now works and does not conflict with the Ruby keyword
 * `ec2_group` validation fixes for `vpc_id`, `purge_rules`, `purge_rules_egress`
 * `lineinfile` module - added missing `validate` parameter
+* `meta` should now work properly like `command` as a free form module
 
 # 1.0.6 - 14 October 2016
 * Allow sharing variables between tasks (#29)